requirements.txt
streamlit
pandas
plotly
fpdf
psycopg2-binary
